nester/requirements.txt
fastapi
uvicorn
pydantic
loguru
pyodbc
python-dotenv
pytest